requirements.txt
# Core Dependencies
flask==3.0.0
transformers==4.36.0
torch==2.1.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# GitHub Integration
PyGithub==2.1.1
GitPython==3.1.40
requests==2.31.0

# Code Processing
tokenize-rt==5.2.0
tree-sitter==0.20.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
